+++ webpack.config.js
@@ -1,7 +1,8 @@
+// @ts-nocheck
+/* eslint-disable */
 const path = require("path");
 const webpack = require("webpack");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
-const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
 
 const devServer = {
 }
@@ -18,11 +19,6 @@ if (process.env.CLOUDIDE_DEV_PORT) {
 module.exports = {
     entry: {
         app: "./src/example.ts",
-        // 'editor.worker': 'monaco-editor/esm/vs/editor/editor.worker.js',
-		// 'json.worker': 'monaco-editor/esm/vs/language/json/json.worker',
-		// 'css.worker': 'monaco-editor/esm/vs/language/css/css.worker',
-		// 'html.worker': 'monaco-editor/esm/vs/language/html/html.worker',
-		// 'ts.worker': 'monaco-editor/esm/vs/language/typescript/ts.worker'
     },
     output: {
         globalObject: 'self',
@@ -60,9 +56,6 @@ module.exports = {
         new HtmlWebpackPlugin({
             template: path.join(__dirname, "src", "index.html"),
         }),
-        new MonacoWebpackPlugin({
-			languages: ['typescript', 'javascript', 'css']
-		}),
         new webpack.ProvidePlugin({
             process: 'process/browser',
         }),